% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_consensus.R
\name{build_consensus_spectrum}
\alias{build_consensus_spectrum}
\title{Build a consensus spectrum from several input spectra.  For spectra derived from the same compound.}
\usage{
build_consensus_spectrum(
  spectra,
  mzTol,
  ppmTol,
  min_rel_int,
  min_rel_rep,
  mz_IQR,
  clean = T,
  min_noise = NULL
)
}
\arguments{
\item{spectra}{a list of spectra in matrix format (mz, intensities)}

\item{mzTol}{numeric tolerance in Daltons}

\item{ppmTol}{numeric tolerance in part per million of mz}

\item{min_rel_int}{numeric cutoff of the relative intensity of the consensus spectrum}

\item{min_rel_rep}{numeric cutoff of the reproducibility of the consensus peak across the spectra}

\item{mz_IQR}{inter-quartile-range of differences of the consensus peak and the corresponding peaks of each spectrum}

\item{clean}{boolean default TRUE  clean the final spectgrum of noise?}

\item{min_noise}{numeric the relative intensity noise floor}
}
\description{
Build a consensus spectrum from several input spectra.  For spectra derived from the same compound.
}
