% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_consensus.R
\name{clean_spectrum}
\alias{clean_spectrum}
\title{Clean spectra of noise}
\usage{
clean_spectrum(spectrum, min_basepeak, min_rel_int, min_noise)
}
\arguments{
\item{spectrum}{matrix the spectrum to be cleaned with mz and intensities (2 columns)}

\item{min_basepeak}{decimal the minimum base peak}

\item{min_rel_int}{the minimum relative intensity to leave in the spectrum}

\item{min_noise}{the min relative intensity noise level (eg .1)}
}
\description{
Clean spectra of noise
}
