% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rawdata.R
\name{computePPMerror}
\alias{computePPMerror}
\title{Calculate the part per million error between a theoretical and real mz value}
\usage{
computePPMerror(mt, me)
}
\arguments{
\item{mt}{numeric theoretical or alternatively measured mz value}

\item{me}{numeric measured mz value}
}
\value{
numeric ppm error
}
\description{
Calculate the part per million error between a theoretical and real mz value
}
