% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentIonData.R
\name{fragment.fgt}
\alias{fragment.fgt}
\title{Fragment ion data store}
\usage{
fragment.fgt(df)
}
\arguments{
\item{df}{fgt data. frame}

\item{collision_type}{collision type eg 'hcd' or 'cid'}

\item{collision_energy}{collision energy (meV) eg '80'}
}
\value{
data frame
}
\description{
Fragment ion data store
}
