% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentIonData.R
\name{fragment.sgd}
\alias{fragment.sgd}
\title{Sgd file dta store}
\usage{
fragment.sgd(df, collision_type = NULL, collision_energy = NULL)
}
\arguments{
\item{df}{data frame from sgd file}

\item{collision_type}{collision type eg 'hcd' or 'cid'}

\item{collision_energy}{collision energy (meV) eg '80'}
}
\value{
data. frame
}
\description{
Sgd file dta store
}
