% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_io.R
\name{generateMS3MSPFile}
\alias{generateMS3MSPFile}
\title{Generate a MSP file for MS3 data}
\usage{
generateMS3MSPFile(spectrum, ct, ce, mz)
}
\arguments{
\item{spectrum}{matrix or numeric 2 columns}

\item{ct}{numeric collision type}

\item{ce}{numeric collision energy}

\item{mz}{numeric m/z}
}
\description{
Generate a MSP file for MS3 data
}
