% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_io.R
\name{generateMSPFile}
\alias{generateMSPFile}
\title{Generate a MSP file for MS2 data}
\usage{
generateMSPFile(spectrum, ct, ce, mz)
}
\arguments{
\item{spectrum}{matrix or numeric 2 columns}

\item{ct}{numeric collision type}

\item{ce}{numeric collision energy}

\item{mz}{numeric m/z}
}
\description{
Generate a MSP file for MS2 data
}
