% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rawdata.R
\name{getMassTolRange}
\alias{getMassTolRange}
\title{Title}
\usage{
getMassTolRange(mass, ppm)
}
\arguments{
\item{mass}{numeric exact mass / mz}

\item{ppm}{numeric ppm tol}
}
\value{
numeric
}
\description{
Title
}
