% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rawdata.R
\name{getMSLevel_Spectra}
\alias{getMSLevel_Spectra}
\title{Get MS level spectra}
\usage{
getMSLevel_Spectra(
  msLevelData,
  msLevel,
  mzTol,
  ppmTol,
  min_rel_int,
  min_rel_rep,
  mz_IQR,
  clean = TRUE,
  min_noise = 0.01
)
}
\arguments{
\item{msLevelData}{list ms Level data from getMzData.  Build consensus spectra and de noises spectra}

\item{msLevel}{integer ms level}

\item{mzTol}{numeric tolerance in Daltons for building consensus spectra}

\item{ppmTol}{numeric tolerance in part per million of mz}

\item{min_rel_int}{numeric cutoff of the relative intensity of the consensus spectrum}

\item{min_rel_rep}{numeric cutoff of the reproducibility of the consensus peak across the spectra}

\item{mz_IQR}{inter-quartile-range of differences of the consensus peak and the corresponding peaks of each spectrum in ppm}

\item{clean}{boolean}

\item{min_noise}{numeric}
}
\value{
list of raw and consensus spectra
}
\description{
Get MS level spectra
}
