% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rawdata.R
\name{getMSLevelData}
\alias{getMSLevelData}
\title{Get ms level data from loaded mz Data using getMzData.}
\usage{
getMSLevelData(mzData, mslevel)
}
\arguments{
\item{mzData}{list from getMzData function.}

\item{mslevel}{integer ms level to extract}
}
\value{
list of data containing index, filter strings, ion tables, collision energies and collision types
}
\description{
Get ms level data from loaded mz Data using getMzData.
}
