% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rawdata.R
\name{getMzData}
\alias{getMzData}
\title{Load mz data using the mzR library.  Returns data in ms levels.}
\usage{
getMzData(mzML_file, mzToRemove = 173.52815, ppmTol)
}
\arguments{
\item{mzML_file}{character vector file name}

\item{mzToRemove}{numeric mz contam to remove}

\item{ppmTol}{numeic ppm tolerance}
}
\value{
list of mzML data segregated into ms levels (eg 1,2,3)  Up to maximum of 3 ms levels.
}
\description{
Load mz data using the mzR library.  Returns data in ms levels.
}
