% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rawdata.R
\name{meanPrecursorMZ}
\alias{meanPrecursorMZ}
\title{Calculate the mean precursor mz from all detected precursor mz in the file. Relevant only for msn data.}
\usage{
meanPrecursorMZ(header, msLevel)
}
\arguments{
\item{header}{mzR header dataframe}

\item{msLevel}{integer ms level (eg, 2, or 3)}
}
\value{
numeric mz value
}
\description{
Calculate the mean precursor mz from all detected precursor mz in the file. Relevant only for msn data.
}
