% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentIonData.R
\name{neutralLoss.fgt}
\alias{neutralLoss.fgt}
\title{Neutral loss data store}
\usage{
neutralLoss.fgt(df, collision_type = NULL, collision_energy = NULL)
}
\arguments{
\item{df}{input fgt data frame}

\item{collision_type}{collision type eg 'hcd' or 'cid'}

\item{collision_energy}{collision energy (meV) eg '80'}
}
\value{
dataframe
}
\description{
Neutral loss data store
}
