% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{normalize.spectrum}
\alias{normalize.spectrum}
\title{normalize the spectrum to 0-100% of the base peak intensity}
\usage{
normalize.spectrum(spectrum)
}
\arguments{
\item{spectrum}{matrix 2 columns}
}
\value{
matrix
}
\description{
normalize the spectrum to 0-100% of the base peak intensity
}
