% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot Spectrum}
\usage{
plot_spectrum(spectrum, name, xlim = c(100, 500))
}
\arguments{
\item{spectrum}{matrix or numeric 2 columns}

\item{name}{character name for plot title}

\item{xlim}{vector x axis limit}
}
\description{
Plot Spectrum
}
