% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_spectrum_match}
\alias{plot_spectrum_match}
\title{Plot a match between 2 spectra}
\usage{
plot_spectrum_match(spectrum, ref_spectrum, xlim = c(100, 500))
}
\arguments{
\item{spectrum}{matrix or numeric 2 columns}

\item{ref_spectrum}{matrix or numeric 2 columns}

\item{xlim}{vector x axis limit}
}
\description{
Plot a match between 2 spectra
}
