% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plotConsensusSpectra}
\alias{plotConsensusSpectra}
\title{Plot a consensus spectrum}
\usage{
plotConsensusSpectra(d, main = "", type = "png", max.range = T)
}
\arguments{
\item{main}{text label for title}

\item{type}{char png or pdf for plot type}

\item{max.range}{bool plot the max range}
}
\description{
Plot a consensus spectrum
}
