% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plotMSnPPMError}
\alias{plotMSnPPMError}
\title{Plot ppm error of ms1 data}
\usage{
plotMSnPPMError(type = "pdf", ppmTol)
}
\arguments{
\item{type}{type 'pdf','png' or device (monitor) plot}

\item{ppmTol}{float ppm tolerance}
}
\description{
Plot ppm error of ms1 data
}
