% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processWorkflow.R
\name{processWorkflow}
\alias{processWorkflow}
\title{process the DNA adduct lbrary building workflow}
\usage{
processWorkflow(
  mzML_file,
  metadata,
  wd = NULL,
  mzTol = 0.005,
  ppmTol = 7,
  min_rel_int = 0.01,
  min_rel_rep = 0.25,
  mz_IQR = 14,
  clean = T,
  min_noise = 0.01,
  contamMZ = 173.52815
)
}
\arguments{
\item{mzML_file}{character mzML file}

\item{metadata}{character array of metadata}

\item{mzTol}{numeric mzTol (Da)}

\item{ppmTol}{numeric ppm tol}

\item{min_rel_int}{numeric rel int}

\item{min_rel_rep}{numeric rel freaction (x/N samples required)}

\item{mz_IQR}{numeric mz IQR of ppm tol to strip outliers by mass error}

\item{clean}{boolean clean the spectra}

\item{min_noise}{noise floor (relative to base peak in consensus)}

\item{contamMZ}{numeric mz value}
}
\description{
process the DNA adduct lbrary building workflow
}
