% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_io.R
\name{read.SGDfile}
\alias{read.SGDfile}
\title{Read sgd file}
\usage{
read.SGDfile(fh)
}
\arguments{
\item{fh}{input sgd file}
}
\value{
data.frame of product ion data and structures
}
\description{
Read sgd file
}
