% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rawdata.R
\name{removeMass}
\alias{removeMass}
\title{Title}
\usage{
removeMass(spectrum, mz, ppmTol)
}
\arguments{
\item{spectrum}{matrix spectrum}

\item{mz}{numeric mz}

\item{ppmTol}{numeric ppm tolerance}
}
\value{
matrix
}
\description{
Title
}
