% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_algo.R
\name{search_library}
\alias{search_library}
\title{Search the library for a spectral match}
\usage{
search_library(spectrum, ref_spectra, N = 10)
}
\arguments{
\item{spectrum}{matrix 2 column}

\item{ref_spectra}{matrix 2 column}

\item{N}{integer N}
}
\description{
Search the library for a spectral match
}
