% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setVars.R
\name{setVars}
\alias{setVars}
\title{Set variables for a new ion}
\usage{
setVars(metadata, wd = NULL)
}
\arguments{
\item{metadata}{character vector of adduct info read from file}

\item{wd}{character user wpecified working dir}
}
\description{
Set variables for a new ion
}
